+++ src/main/java/gtPlusPlus/xmod/gregtech/common/tileentities/machines/multi/processing/GregtechMetaTileEntity_IndustrialForgeHammer.java
@@ -171,8 +171,8 @@ public class GregtechMetaTileEntity_IndustrialForgeHammer extends
     }
 
     @Override
-    public String getSound() {
-        return SoundResource.RANDOM_ANVIL_USE.toString();
+    protected SoundResource getProcessStartSound() {
+        return SoundResource.RANDOM_ANVIL_USE;
     }
 
     @Override
