+++ src/main/java/gtPlusPlus/xmod/gregtech/common/tileentities/machines/multi/processing/advanced/GregtechMetaTileEntity_Adv_Implosion.java
@@ -157,8 +157,8 @@ public class GregtechMetaTileEntity_Adv_Implosion
     }
 
     @Override
-    public String getSound() {
-        return SoundResource.RANDOM_EXPLODE.toString();
+    protected SoundResource getProcessStartSound() {
+        return SoundResource.RANDOM_EXPLODE;
     }
 
     @Override
